refactor(scenarios): rename scenario-run.service to scenario-run.utils (#2049)
The file exports pure helper functions, not a service class. Rename to
follow the existing `<domain>.<concern>.ts` convention.
